[#2449] Generalized code from Action and Commands
Extracted common code to Abstract class

Also added missing .project file for RCPTT
