+++ core/plugins/org.polarsys.capella.core.sirius.analysis/META-INF/MANIFEST.MF
@@ -12,6 +12,7 @@ Export-Package: org.polarsys.capella.core.sequencediag,
  org.polarsys.capella.core.sirius.analysis.actions.extensions,
  org.polarsys.capella.core.sirius.analysis.activator,
  org.polarsys.capella.core.sirius.analysis.cache,
+ org.polarsys.capella.core.sirius.analysis.commands,
  org.polarsys.capella.core.sirius.analysis.constants,
  org.polarsys.capella.core.sirius.analysis.delete,
  org.polarsys.capella.core.sirius.analysis.editpart,
