+++ core/plugins/org.polarsys.capella.core.sirius.ui/src/org/polarsys/capella/core/sirius/ui/actions/NewRepresentationAction.java
@@ -19,18 +19,11 @@ import org.eclipse.jface.dialogs.IInputValidator;
 import org.eclipse.jface.dialogs.InputDialog;
 import org.eclipse.jface.resource.ImageDescriptor;
 import org.eclipse.jface.window.Window;
-import org.eclipse.sirius.business.api.query.IdentifiedElementQuery;
 import org.eclipse.sirius.business.api.session.Session;
 import org.eclipse.sirius.business.api.session.SessionManager;
 import org.eclipse.sirius.common.tools.api.interpreter.EvaluationException;
 import org.eclipse.sirius.common.tools.api.interpreter.IInterpreter;
-import org.eclipse.sirius.common.tools.api.util.MessageTranslator;
 import org.eclipse.sirius.common.tools.api.util.StringUtil;
-import org.eclipse.sirius.diagram.sequence.description.SequenceDiagramDescription;
-import org.eclipse.sirius.diagram.ui.provider.DiagramUIPlugin;
-import org.eclipse.sirius.table.metamodel.table.description.CrossTableDescription;
-import org.eclipse.sirius.table.metamodel.table.description.EditionTableDescription;
-import org.eclipse.sirius.table.metamodel.table.provider.TableUIPlugin;
 import org.eclipse.sirius.tools.api.SiriusPlugin;
 import org.eclipse.sirius.tools.api.interpreter.IInterpreterMessages;
 import org.eclipse.sirius.tools.api.interpreter.InterpreterUtil;
@@ -38,25 +31,19 @@ import org.eclipse.sirius.ui.business.api.dialect.DialectUIManager;
 import org.eclipse.sirius.viewpoint.description.RepresentationDescription;
 import org.eclipse.swt.widgets.Display;
 import org.eclipse.swt.widgets.Shell;
-import org.eclipse.ui.actions.BaseSelectionListenerAction;
-import org.eclipse.ui.plugin.AbstractUIPlugin;
 import org.polarsys.capella.common.helpers.TransactionHelper;
 import org.polarsys.capella.common.ui.toolkit.dialogs.Messages;
-import org.polarsys.capella.core.sirius.analysis.NewRepresentationCommand;
+import org.polarsys.capella.core.sirius.analysis.commands.NewRepresentationCommand;
 
 /**
  * The action allowing to create new representations.
  */
-public class NewRepresentationAction extends BaseSelectionListenerAction {
-	private EObject selectedEObject;
+public class NewRepresentationAction extends AbstractNewRepresentationAction {
+	
 	protected RepresentationDescription description;
-	protected Session session;
-
-	protected boolean forceDefaultName;
-	protected boolean openRepresentation;
-	private boolean isCanceled;
+	
+	protected boolean forceDefaultName;	
 	private String descriptionLabel;
-	private String message;
 
 	public NewRepresentationAction(RepresentationDescription description, EObject selectedEObject, Session session, String message) {
 		this(description, selectedEObject, session, false, true);
@@ -84,37 +71,6 @@ public class NewRepresentationAction extends BaseSelectionListenerAction {
 
 		ImageDescriptor imageDescriptor = getDescriptionImageDescriptor(description);
 		setImageDescriptor(imageDescriptor);
-	}
-
-	protected String getDescriptionLabel(RepresentationDescription description) {
-		return MessageTranslator.INSTANCE.getMessage(description, new IdentifiedElementQuery(description).getLabel());
-	}
-
-	protected ImageDescriptor getDescriptionImageDescriptor(RepresentationDescription description) {
-		ImageDescriptor imageDescriptor = null;
-
-		// Handle specific representations : Table ones.
-		if (description instanceof CrossTableDescription) {
-			imageDescriptor = AbstractUIPlugin.imageDescriptorFromPlugin(TableUIPlugin.ID,
-					"/icons/full/obj16/CrossTableDescription.gif"); //$NON-NLS-1$
-
-		} else if (description instanceof EditionTableDescription) {
-			imageDescriptor = AbstractUIPlugin.imageDescriptorFromPlugin(TableUIPlugin.ID, "/icons/full/obj16/DTable.gif"); //$NON-NLS-1$
-
-		} else if (description instanceof SequenceDiagramDescription) {
-			imageDescriptor = AbstractUIPlugin.imageDescriptorFromPlugin("org.eclipse.sirius.diagram.sequence.edit", //$NON-NLS-1$
-					"/icons/full/obj16/TSequenceDiagram.gif"); //$NON-NLS-1$
-
-		} else {
-			// Standard diagram.
-			imageDescriptor = AbstractUIPlugin.imageDescriptorFromPlugin(DiagramUIPlugin.ID,
-					"/icons/full/obj16/DDiagram.gif"); //$NON-NLS-1$
-		}
-
-		if (null == imageDescriptor) {
-			imageDescriptor = ImageDescriptor.getMissingImageDescriptor();
-		}
-		return imageDescriptor;
 	}
 
 	protected String computeDefaultName(EObject eObject, RepresentationDescription repDescription) {
@@ -190,15 +146,5 @@ public class NewRepresentationAction extends BaseSelectionListenerAction {
 				DialectUIManager.INSTANCE.openEditor(session, command.getRepresentation(), new NullProgressMonitor());
 			}
 		}
-	}
-
-
-
-	public boolean isCanceled() {
-		return isCanceled;
-	}
-
-	public void setCanceled(boolean isCanceled) {
-		this.isCanceled = isCanceled;
 	}
 }
