+++ core/plugins/org.polarsys.capella.core.sirius.ui/src/org/polarsys/capella/core/sirius/ui/actions/NewScenarioRepresentationAction.java
@@ -16,24 +16,17 @@ package org.polarsys.capella.core.sirius.ui.actions;
 import org.eclipse.core.runtime.NullProgressMonitor;
 import org.eclipse.jface.dialogs.InputDialog;
 import org.eclipse.jface.window.Window;
-import org.eclipse.sirius.business.api.dialect.DialectManager;
 import org.eclipse.sirius.business.api.query.IdentifiedElementQuery;
 import org.eclipse.sirius.business.api.session.Session;
 import org.eclipse.sirius.business.api.session.SessionManager;
 import org.eclipse.sirius.common.tools.api.util.MessageTranslator;
 import org.eclipse.sirius.ui.business.api.dialect.DialectUIManager;
-import org.eclipse.sirius.viewpoint.DRepresentation;
 import org.eclipse.sirius.viewpoint.description.RepresentationDescription;
 import org.eclipse.swt.widgets.Display;
 import org.eclipse.swt.widgets.Shell;
-import org.polarsys.capella.common.ef.command.AbstractReadWriteCommand;
 import org.polarsys.capella.common.helpers.TransactionHelper;
-import org.polarsys.capella.common.tools.report.appenders.usage.UsageMonitoringLogger;
-import org.polarsys.capella.common.tools.report.appenders.usage.util.UsageMonitoring.EventStatus;
 import org.polarsys.capella.core.data.interaction.AbstractCapability;
-import org.polarsys.capella.core.data.interaction.InteractionFactory;
-import org.polarsys.capella.core.data.interaction.Scenario;
-import org.polarsys.capella.shared.id.handler.IdManager;
+import org.polarsys.capella.core.sirius.analysis.commands.NewScenarioRepresentationCommand;
 
 /**
  * The action allowing to create new representations.
@@ -103,71 +96,5 @@ public class NewScenarioRepresentationAction extends NewRepresentationAction {
       }
     }
   }
-
-  // The command allowing to create a new representation.
-  private class NewScenarioRepresentationCommand extends AbstractReadWriteCommand {
-    // The representation name.
-    private String newName;
-    // The new representation.
-    private DRepresentation representation;
-
-    // Fields.
-    private RepresentationDescription repDescription;
-    private Session currentSession;
-
-    /**
-     * Constructs the command allowing to create a new representation.
-     * @param newName The new representation name.
-     * @param eObject The selected EObject.
-     * @param repDescription The current representation description.
-     * @param session The current session.
-     */
-    public NewScenarioRepresentationCommand(String newName, AbstractCapability eObject, RepresentationDescription repDescription, Session session) {
-      this.newName = newName;
-      this.repDescription = repDescription;
-      this.currentSession = session;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void commandInterrupted() {
-      commandRolledBack();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void commandRolledBack() {
-      representation = null;
-    }
-
-    /**
-     * Gets the new representation.
-     * @return The new representation.
-     */
-    public DRepresentation getRepresentation() {
-      return representation;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @SuppressWarnings("synthetic-access")
-    public void run() {
-      NullProgressMonitor monitor = new NullProgressMonitor();
-      Scenario scenario = InteractionFactory.eINSTANCE.createScenario();
-      scenario.setName(newName);
-      selectedEObject.getOwnedScenarios().add(scenario);
-
-      String eventName = "Create Representation";
-      String eventContext = repDescription.getName();
-      String addendum = IdManager.getInstance().getId(scenario);
-      UsageMonitoringLogger.getInstance().log(eventName, eventContext, EventStatus.NONE, addendum);
-      representation = DialectManager.INSTANCE.createRepresentation(newName, scenario, repDescription, currentSession, monitor);
-      UsageMonitoringLogger.getInstance().log(eventName, eventContext, EventStatus.OK, addendum);
-    }
-  }
+  
 }
