+++ core/plugins/org.polarsys.capella.core.sirius.ui/src/org/polarsys/capella/core/sirius/ui/actions/messages.properties
@@ -1,0 +1,14 @@
+#===============================================================================
+# Copyright (c) 2006, 2022 THALES GLOBAL SERVICES.
+# 
+# This program and the accompanying materials are made available under the
+# terms of the Eclipse Public License 2.0 which is available at
+# http://www.eclipse.org/legal/epl-2.0
+# 
+# SPDX-License-Identifier: EPL-2.0
+# 
+# Contributors:
+#    Thales - initial API and implementation
+#===============================================================================
+
+SelectNewRepresentation=Select a representation to create
